Solve not using custom error message
This is the function in which is responsible for deciding which error message to use

```javascript
// src/shared/lib/forms/validationAdapters/joi.js

/**
 * Formats error object from Joi into an easy format, and includes
 * custom error messages from the object provided
 * @param {Object} error - Joi error from schema.validate()
 * @param {Object} customErrors - custom error messages
 * @return {Array} formatted error messages
 */
const formatErrors = (error, customErrors) => {
  const details = get(error, 'details', [])

  const errors = details.map(err => {
    const name = err.context.key
    const { type, message } = err

    // Use custom error messages
    if ((name in customErrors) && (type in customErrors[name])) {
      const custom = customErrors[name][type]
      return {
        name,
        message: custom.message,
        summary: custom.summary || custom.message
      }
    }

    // Use default Joi message
    return {
      name,
      message,
      summary: message
    }
  })

  // only return one error for each field
  return uniqBy(errors, 'name')
}
```

The reason it is failing for us is because the custom error we apply in `src/internal/modules/address-entry/forms/select-address.js form()` has the type `'any.only'`. This doesn't match what Joi is returning when it validates the UPRN which is `'number.base'`.

So, no match is found and the function resorts to using Joi's error message.