Refactor and add documentation
Whilst trying to debug the issue we spotted that we were calling `formContainer.form(request)` multiple times to get the same result. So, first change was to avoid repeating the call unnecessarily.

But we also had to deep dive into what on earth this was supposed to be doing. So, we have also added some documentation that hopefully will still make sense should we return to this section of the code!