Fix for the empty lookup list
The key find was the fact that after a post request if there is a validation error they stash the form object in the session and then retrieve it wherever they send the redirect request to.

That logic depended on the form object also having the addresses set in the form's address dropdown field. This was how it could re-show them again. But with us removing the pre-handler till after we have validated a selection has been made we broke this process.

So, our 'hack' is to re-populate the field in the GET controller when we find a form object with errors.