Fix no address selected 500 error (#2315)
https://eaflood.atlassian.net/browse/WATER-3943

This was spotted when attempting to sign off the next release of the service and running through creating a new billing account with a new address during charge version creation. Having entered a postcode and searched for it you are presented with the drop-down of addresses found from which you must make a selection.

If you don't select anything and click continue, the validation message should be 'Select an address from the list'. Instead, the Ui is erroring and returning a `500` to the user.

This is not the result of a recent change, but one we made back in August 2022; [Remove catbox-redis](https://github.com/DEFRA/water-abstraction-service/pull/1794) from service So, the error has been present for some time. The reason given for adding [catbox-redis](https://github.com/hapijs/catbox-redis) by the previous team was

> caching avoids API rate limit being exceeded on subsequent identical requests

As part of our general maintenance of the service, this was identified as being woefully out-of-date and was going to need us to re-write our code in order to work with the latest version.

But we also knew that we are using a version of the [Address-facade](https://github.com/DEFRA/ea-address-facade) which has a rate limiter built-in so we are protected from getting into trouble with OS Places. And we are not a service that sees enough traffic to be hitting their rate limiter anyway. So, this appeared to be a performance/protection change that was unnecessary. On that basis, we instead chose to remove **catbox-redis**. One less dependency, code deleted and a baby step towards a simpler solution.

When we investigated the problem we found who was causing us to hit the rate limit by performing the equivalent of a DOS attack; us!

The sequence we found is

- user submits postcode and UI redirects the request to the address selection page after validating the postcode
- redirect GET request received by UI and pre-handler function calls the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) to perform the postcode search
  - address-facade receives request, hits OS Places and returns results to the water-abstraction-service
  - water-abstraction-service returns the results to the UI
- UI returns the page to the user with the address dropdown populated with the results (UPRN and address for each one)
- user clicks continue without selecting an address
- POST request received by UI and pre-handler function calls the water-abstraction-service to perform the postcode search (it is grabbing the results in preparation for identifying which one has the UPRN the user selected)
- UI deems the request invalid because no UPRN was selected and redirects to the address selection page
- redirect GET request received by the UI and the pre-handler function calls the water-abstraction-service to perform the postcode search
  - address-facade receives request and rejects it. This is because the last 2 requests come within a few milliseconds of each other
  - water-abstraction-service does not handle the error so returns a 500 to the UI
- UI does not handle the error so returns a 500 to the user

It's the previous team's love of [pre-handlers](https://hapi.dev/api/?v=21.3.0#route.options.pre) to blame here. We shouldn't be making the request for address results in the POST request until we have verified that the user has submitted a UPRN. But the previous team liked to hide a lot of functionality in these pre-handler methods which are always called before the controller. It's very much a "we have a groovy new hammer. Now everything looks like a nail" approach.

This change does what we think should be done; drop the pre-handler on the POST request in the UI and only call the water-abstraction-service once we have confirmed a UPRN has been selected.

But nothing is ever simple! This area of the UI is where the previous team decided to use the custom form engine they built into the app! So, all the pages use generic code to generate the page based on properties, modules and functions you have to provide.

In the case of the select address 'form' they have not been consistent in their approach. Some places protect against the address results being blank. In others it is assumed you have results and that the pre-handler has run. So, without them, we cause other errors to be thrown.

We also found that the error [joi](https://github.com/hapijs/joi) generates when the UPRN is missing doesn't match the custom error assigned to this custom field in the custom form. I mean, for the love of...! 😩🤦

So, we fix all that as well 😁