Update package `file-type` and remove `read-chunk`
https://github.com/DEFRA/water-abstraction-team/issues/78

After working on issue Swap out or remove Deprecated packages for the water-abstraction-ui in this PR #2302 we found that in order to update/replace/remove some of the packages a significant amount of work was needed. These are therefore going to be done individually in separate PRs.

In this PR the package `file-type` is being updated. `read-chunk` can be removed as after checking the code we found that`file-type` was being used to check the file type by first reading in the file using `read-chunk`. File-type was then used to check the file type from the buffer and return just the file extension as a text string. We found that this is unnessessary and that there is no benifit to determining the file type in this way so the `file-type` method of `fileTypeFromFile` is now used to save us using `read-chunk` to read the file into the buffer first.