Implement change
We reduce the logic in the view to just be 'if thing exists show it'. TBH, that should be the limit of logic in a view.

We move the logic of determining what to show into the controller by adding a new function, used to set a new property in the view data.

We expand the tests to include coverage for the changes we are making.