Update include licence in supp. billing messages
https://eaflood.atlassian.net/browse/WATER-3948

In [Highlight flagged for SROC supplementary billing](https://github.com/DEFRA/water-abstraction-ui/pull/2322) we made a change to the notice shown to users when a licence is flagged for supplementary billing. The first iteration of the wording was what us developers had put together based on the existing messages and changes elsewhere in the service.

Our UAT team have now had a chance to look at them and asked for some tweaks.