Addional charges incorrect for supported sources
https://eaflood.atlassian.net/browse/WATER-4001

Following the fix in https://github.com/DEFRA/water-abstraction-service/pull/2142. Which fixed the issue where the incorrect value was being used for the Support Sources additional charges.

It has been found during testing that this value is incorrectly being formatted from pence to pounds. Since the corrected value is already in pounds and needs no conversiobn.

This PR will fix that issue.