Update ESM only package `parse-domain`
https://github.com/DEFRA/water-abstraction-team/issues/78

The package `parse-domain` is 2 major versions out of date. Unfortunatly since version 6 this package is ESM only.

To import this package asyncronously using CommonJS. This has been a bit of work as the function in which `parse-domain` is required is syncronous so several things needed to be fixed to work asyncronously.