Add SROC only billing feature flag
https://eaflood.atlassian.net/browse/WATER-4016

The billing & data team are struggling with the performance of the service in PREPROD during UAT. In WATER-4013 we were able to determine that because of the poor quality of queries in the legacy code the service is being throttled by AWS because we're hitting the IOPS threshold.

We've made changes to help with that in PRE-PROD (doubled the size of the RDS instance!) but we are still faced with a service that becomes overwhelmed if you try to run more than one bill run at a time. Until we've been able to migrate away from the legacy code we are always going to have performance issues. It is why the current dev team's guidance is only one bill run at a time (followed by a nice brew, some friendly chat, and generally avoid touching it for awhile!)

But the billing & data team want to get on with their UAT. Different people look after different regions and they have a lot of scenarios they need to work through.

So, we've had a brain wave. As the focus is on the SROC bill runs, what if we switched off the PRESROC ones temporarily?

This change covers adding a feature flag to the service, which if enabled would mean the service will only kick off the SROC bill run.