Remove package `parse-domain` (#2354)
https://github.com/DEFRA/water-abstraction-team/issues/78

The package `parse-domain` is 2 major versions out of date. Unfortunately, since version 6 this package is ESM only.

To import this package asynchronously using CommonJS is a bit of work as the function in which `parse-domain` is required is synchronous so several things would need to be updated for it to work asynchronously.

Fortunately though from going through the code we have found that it isn't really doing anything much apart from removing the subdomain from the `hostName` if it exists. Since the names of our domains are known we don't need this package to work them out for us.