Warn the user if a bill run is 'processing' (#2369)
https://eaflood.atlassian.net/browse/WATER-4030

The B&D frequently report issues with performance in the WRLS service. The same goes for errors when running bill runs.

This isn’t surprising to the new dev team based on how we now understand the service has been built and architected.

Since joining, we have advocated that the B&D should only run one bill at a time. We believe this will reduce the performance hit on the overall service and the likelihood of errors.

It seems that this message has not always been received or understood. Recent supplementary billing testing highlighted this when all the region leads were trying to test bill runs simultaneously.

This PR adds a notification banner to the 'Bill runs' page which is displayed if a bill run has a status of 'Building' or 'Queued'. This notification instructs the user to wait for the bill run to finish before creating another one.