Fix name used for billing contact
https://eaflood.atlassian.net/browse/WATER-3973

> TL;DR show the same name we send to the [Charging Module API's customer-changes endpoint](https://defra.github.io/sroc-charging-module-api-docs/#/customer/CreateCustomerChange).

-----

As part of creating a new billing account you will select a 'customer (company)' and an address. Assume you selected the same 'company' as the previous charge version (the address doesn't matter).

You then get the option to create an 'FAO'. When you do you are creating a crm_v2.contacts record. This will automatically be linked to the company you selected to use for the new invoice account via a `crm_v2.company_contacts` record.

It will also create a new `crm_v2.invoice_account_addresses` record which links the address and new contact records to the `crm_invoice_account` we're creating.

This is the key difference between an invoice account with an FAO and one without; the `contact_id` in the `crm_v2.invoice_account_addresses` record will be empty.

So, onto the problem. When displaying the invoices the one linked to an invoice account with an FAO (a contact) displays as expected.

Any without a contact falls back on other logic. The service goes back to the company record and looks for any contacts linked to it with the role of 'Billing'. And we now know the current service behaviour is to link any new FAO contact to the company. That logic finds the FAO contact again, so displays the name again.

This is the critical bit of code in the `src/internal/views/nunjucks/billing/macros/invoices-table.njk`

```html
{% macro billingContactCell(invoice) %}
  <td class="govuk-table__cell">
    {% if invoice.faoContact  %}
      {{ invoice.faoContact }}

      {% elif invoice.billingContact.agentCompanyName %}
      {{ invoice.billingContact.agentCompanyName }}

      {% elif invoice.billingContact.roleContact.licenceHolder %}
      {{  invoice.billingContact.roleContact.licenceHolder }}

      {% elif invoice.billingContact.roleContact.billing %}
      {{  invoice.billingContact.roleContact.billing }}

    {% else %}
      {{ invoice.name | titleCase }}
    {% endif %}

  </td>
{% endmacro %}
```

Before the charge version with the FAO was added (and assuming no previous contact exists for the company) it would fall into the final if/else block and just display the company name.

When the charge version with an FAO is created it results in `elif invoice.billingContact.roleContact.billing` returning a value for the invoice linked to the charge version without an FAO.

IMHO, the CRM entities have been so over complicated it's no wonder you see behaviour like this. But as it stands, its been specifically coded to behave this way, i.e.

- if the invoice account address has a contact use it
- else if the invoice account company has any contact, use the latest (oh yeah, add another contact and you'll see the latest displayed instead!)
- else just display the company name

-----

That's the background for this change. When we went back to the users for clarification we got a very clear direction back; use the same name that is sent to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). When an invoice account is created or amended we have to let the Charging Module know. This is because it generates a 'customers' file that is sent to SOP for use when generating the invoices. That logic is different to the UI hence the discrepency.

This change updates the UI logic to match what [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) is doing when generating the `/customer-changes` request to the CHA.