Amend logic in the bill runs view
Bring inline with the water-abstraction-service. It uses the name of the agency company linked to the latest invoice address if it exists, else the company directly assigned to the invoice account.

It appears the UI has already determined the agency company. So, if it exists we use it else the invoice name.