Implement fix for bill run invoice summary screen
The water-abstraction-service uses models and has a function on the `InvoiceAccount` model to determine the latest invoice address.

We're not so lucky and have to work with what we've got. We add a method that performs the same logic the service does, determine the latest address and use the agency company name linked to it if it exists.

Else just use the name of the company directly linked to the invoice account.