Fix logic for tests
A number of tests failed because they were not providing invoice accounts with addresses. As far as we understand this is an impossibility. But rather than go through all the broken tests and ensuring they setup in such a way addresses are present, we update the logic to handle this scenario.