Rename existing System connector service
On the cusp of adding a new connector for the `/billing-accounts` endpoint we're adding we realised that the existing `BillRunsService` didn't follow the naming convention of the _majority_ of the other connector services.

Renaming it to `BillRunsApiClient.js` not only makes it more consistent it also makes it clearer what kind of connector it is.