Redirect /service-status to system app (#2426)
https://eaflood.atlassian.net/browse/WATER-4099

We are replacing the existing `/service-status` page with an updated version. But as all new code is being developed in [water-abstraction-system](https://github.com/water-abstraction-system) we need the UI to redirect the request to there.

This change just covers that. We'll look at what code we can delete when we are happy everything is working as expected.