Fix service status redirect to system
https://eaflood.atlassian.net/browse/WATER-4099

In [Redirect /service-status to system app](https://github.com/DEFRA/water-abstraction-ui/pull/2426) we redirected the `/service-status` endpoint to `/health/info` in the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) app.

But we were too clever and thought we needed the address for the system app. It works locally but in our AWS environments the redirect would fail because you cannot access the system app directly.

So, instead we need to redirect the request through the proxy with implemented within the UI which passes requests onto our system app.