Generate unique job ID for notifications (#2451)
https://eaflood.atlassian.net/browse/WATER-4110

As part of our effort to prevent unintentional duplication of 'hands-off flow level warnings', this PR focuses on adding a unique identifier during the process. Within our notification system, there exist four distinct types of notifications, and one of these involves the 'hands-off flow warning letters'. To avoid the scenario where multiple identical requests are processed resulting in the dispatch of multiple letters or emails, the work presented in this PR aims to generate a unique job ID during the preview phase of the notifications flow.
This distinctive job ID will then be sent to the "sendNotifications" job. Our system utilizes BullMQ for managing queues, workers, and jobs. BullMQ is configured in such a way that if a job with a unique ID is added to the queue, any subsequent job added with the same ID will be disregarded to prevent duplicate execution. By employing this unique ID as the job ID, we ensure that the job is not executed more than once, effectively addressing the issue of duplicate job processing.