Generate unique job ID for paper return forms
https://eaflood.atlassian.net/browse/WATER-4106

As part of our effort to prevent unintentional duplication of 'paper return forms', this PR focuses on adding a unique identifier during the process.
To avoid multiple identical requests from being processed (resulting in the dispatch of multiple letters), the work presented in this PR aims to generate a unique job ID during the preview phase of the paper returns flow. This job ID will then be sent to the "postPrepare" job in the service. we are then using this for a check in the service to see if the request has already happened.