Link to new bill page from bill run summary
We had two ways we could get the UI to start using the new page.

We could have left the links as they were and have the receiving controller then redirect the request to **water-abstraction-system**. The benefit of that approach is we just have to make a change in one place; the controller. But it does require a redirect and lost time whilst the browser and the servers deal with this.

The alternative is to change the links we are generating so that when clicked they go straight to our new page. No redirect required but due to the _insanity_ of the previous teams macro usage it does feel like we have to change every template to achieve it!

But we like a challenge! Plus we have something to prove when it comes to performance. So, endless macro updating it is.