Link to new page from billing account x2!
The acceptance criteria on the ticket requires us to link to the new bill page from both the bill run summary and the billing account page.

What it didn't clock was that if a billing account has lots of bills, at link at the bottom of the page appears to view them all. So, these changes cover updating the links in both places. Phew!