Update links in missed source (licence view)
Our mistake this one as it was in the ticket 🤦. You can also link to a bill via the 'Bills' tab in the view licence page.

The control is already passing in the whole `featureToggles` object. So, we just needed to make sure our `useNewBillView` flag made it down to the macro that is generating the link.