Link to new bill page (#2461)
https://eaflood.atlassian.net/browse/WATER-4155

We are working on implementing a new version of the bill page. The current one tries to display all transaction details for all licences on a single page when the bill is selected.

In the larger annual bill runs this might equate to more than 150 licences and their transactions which means the page is failing to load.

All that work is being done in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) as part of our ongoing efforts to migrate away from the legacy code.

This means we need to update the UI to point to the system app and the endpoint we are building there when a user follows a view bill link.

We know there are 3 places we link to a bill; from the bill run summary, when looking at the bills for a selected billing account and from the licence summary 'Bills' tab. Because the work is not complete we'll put the change behind a feature flag to allow us to ship other changes if we need to.