Prevent SROC 2PT being triggered when FYE < 2023
https://eaflood.atlassian.net/browse/WATER-4197

Following on from the work we did to create a trigger for the SROC two-part-tariff bill runs we have realised that the bill run can be triggered for billing periods outside of the SROC billing period.

This PR will fix that issue by preventing an SROC 2PT bill run from being triggered when the `financialYearEnding` is prior to 2023.