Link to new bill page from all sent bills page
https://eaflood.atlassian.net/browse/WATER-4131

We have been working on replacing the existing views of a bill in the legacy code because they are slow and timeout when attempting to view large bills with many licences.

In [Link to new bill page](https://github.com/DEFRA/water-abstraction-ui/pull/2461) we thought we had updated all existing links that point to a bill to instead point to our new page.

But we missed one. After searching for a licence and having selected the 'Bills' tab there is a link at the bottom that takes you to a 'View all sent bills page'. If you click on a bill in there it takes you to the legacy view.

This change updates that link as well to point to the new bill view.