Implement change
The template is already using a macro we edited to enable the new link to be shown. But the controller and template were missed so we didn't make the change to pass the feature toggle from the controller to the macro via the template.

That's all we've had to do to make this work.