Fix broken unit test
Removing the unnecessary check on batch type when determining whether to redirect to the new SROC review page has highlighted that the test is wrong. It was passing before because the type the test is setting is what you see in the UI not what you get from the DB.

So, before the change in this PR it wasn't doing what you think it should be doing i.e. generating a link to the **water-abstraction-system** because `Two-part tariff` does not equal `two_part_tariff`. Having removed the check on type it is now generating the link, but the test is asserting you get the old ALCS one back!