Make bill run summary page back link fixed
https://eaflood.atlassian.net/browse/WATER-4131

The new bill pages we have been developing to replace the existing ones take a different approach to backlinks. Currently, in the service backlinks are intended to act like the browser back button. Clicking them should take you back to wherever you came from.

That's not really the intent of them in the design system though. When added to a page they _should_ take you back to the step or context before where you currently are. So, if you are on the bill page, you should go back to the bill run that bill is linked to.

The browser back button is there for those who truly want to go back to wherever they came from.

Another reason for following this guidance is backlinks that just trigger `javascript:history.back()` are notorious for getting people stuck in loops. And that is now happening with our new pages. Suppose a user comes to the new bill page via the bill run summary page. Our new fixed link in will take them back to the bill run summary. But the back link in that page will just trigger the javascript sending them back ... to the bill page!

To avoid this and at the same time bring the page in line with our new ones, this change updates the backlink to always go back to the bill runs list page.