Fix back link
Always show the backlink and have it always go back to the bill runs screen. Our new back links use text to tell the user where they are going back to. In a journey `< back` is sufficient because the user will know there they are going back to. But in pages where they may have come from different locations it is important to state where the link will take you back to.