Redirect to new bill run page (#2473)
https://eaflood.atlassian.net/browse/WATER-4223

We have been working on implementing a new version of the bill page. The current one tries to display all transaction details for all licences on a single page when the bill is selected.

In the larger annual bill runs this might equate to more than 150 licences and their transactions which means the page is failing to load.

All that work has been done in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) as part of our ongoing efforts to migrate away from the legacy code. It is in the final throws of testing and will soon be released.

To get to the new pages you have to pass through the existing bill run summary page. What our work has highlighted is this page also leaves a lot to desire! Again, the larger the bill run the slower it is, noticeably so. Though it at least always loads 😁. It also is inconsistent. It has a different layout depending on the bill run type, and one of those depends on tabs which is an incorrect use of the component.

We'd like to remove this remaining 'bump' in our billing user journey so have rebuilt the page in **water-abstraction-system**. So, now if the feature toggle `USE_NEW_BILL_VIEW` is enabled we direct the user to the new page.

---

We also use the opportunity to simplify what is going on in `getBillingBatchRoute()`. At least we think it is simpler and clearer!

Plus we fix a broken unit test in `test/internal/modules/billing/controllers/bill-run.test.js`. Removing the unnecessary check on batch type when determining whether to redirect to the new SROC review page highlighted that the test was wrong. It was passing because the batch type the test was setting is what you see in the UI not what you get from the DB.

So, before this change, it wasn't doing what you think it should be doing i.e. generating a link to the **water-abstraction-system** because `Two-part tariff` does not equal `two_part_tariff`. This meant it generated a link to the ALCS review page which is fine because the test was also asserting that's where you go to!

Having removed the batch type check from the code it is now generating a link to our new review page and asserting this.