Update controller to use feature flag
The difference between annual and supplementary is that we'll only ever request one bill run. Only SROC applies from 2022 and we only generate Annual bill runs for the current year. So, the only reason to send a request to the water-abstraction-service is of the feature toggle is false.

So, we now need to use the response from water-abstraction-system to obtain the newly created billing batch ID.