Enable proxying to system from external UI
https://eaflood.atlassian.net/browse/WATER-4352

The legacy UI is one code base split into 3; internal, external and shared.
When deployed we have 2 copies of the code base running; one acts as the internal app and the other the external facing.
We have already implemented proxying to water-abstraction-system in the internal code.
This ticket is about moving that code to "shared" so it can be used by both apps.