Fixing proxying to system from external UI
https://eaflood.atlassian.net/browse/WATER-4352

https://github.com/DEFRA/water-abstraction-ui/pull/2507

When trying to access the external site, found there was an issue causing "This site can't be reached" being displayed.

Upon inspection, this was being caused by a missing import for proxying configuration within the external which was already present in the internal UI.

This change adds the configuration needed for the external site to work.