Remove use new bill run views feature flag
https://eaflood.atlassian.net/browse/WATER-4397

What started out as being an issue displaying large bills turned into a complete re-write of the view bill run pages. Whilst we worked on the re-design we put the new pages behind a feature flag.

Those new pages were shipped more than a month ago and are now the norm. There is no need to retain the feature flag. Removing it makes the code simpler. This change does that.