Remove use new bill run views feature flag (#2523)
https://eaflood.atlassian.net/browse/WATER-4397

What started as an issue displaying large bills turned into a complete rewrite of the view bill run pages. Whilst we worked on the re-design we put the new pages behind a feature flag.

Those new pages were shipped more than a month ago and are now the norm. There is no need to retain the feature flag. Removing it makes the code simpler. This change does that.