Link errored bill runs to view bill page (#2525)
https://eaflood.atlassian.net/browse/WATER-4398

We have replaced the legacy version of the view errored bill run page with one in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

This updates the routing logic to return a link to the new view. If you look at the change you'll see this is the same route as when viewing a ready or sent bill run. In our new repo, we don't make separate routes based on status. Instead, it is one `GET /bill-runs/{id}` route to view a bill run and logic in the app will decide how the bill run is presented.