Display cancelling bill runs in bill runs page
https://eaflood.atlassian.net/browse/WATER-4387

We need to add this to support our work to make cancelling bill runs visible to users.

Cancelling a bill run, especially a large one, can impact the service and take a considerable amount of time (30 minutes for the largest ones!) We want to let users know when a bill run is being cancelled so that they can make a more informed decision as to when it is best to try creating a new bill run.

We've amended [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service/pull/2450) to stop filtering bill runs with a status of `cancel` from the bill run results. The final step is to update this project to ensure they are displayed correctly in the table. This means they have a badge and no link to view the bill run.