Update cancel btn in 2pt review to new process (#2527)
https://eaflood.atlassian.net/browse/WATER-4387

In [Display cancelling bill runs in bill runs page](https://github.com/DEFRA/water-abstraction-ui/pull/2526) we made changes to help make cancelling bill runs more visible.

This was all built on the work we'd done to replace the legacy cancel bill run process with a new one in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

We overlooked that some pages in the legacy UI still need their cancel buttons re-pointed to the new endpoint. It was the PRESROC two-part tariff review page which caught our eye. But as we're updating a macro it'll mean any page that uses it will also benefit.